Sign an integer timestamp, not a float
The announcement carried t as an f64. serde_json writes 1787862155.6581013
and reads it back as 1787862155.658101, so the verifier canonicalized
different bytes than the sender signed and the HMAC failed. It failed for
some values and not others, so discovery half worked: roughly every other
announcement was dropped, and a daemon still reached the watch set through
the ones that survived.

spark-agent sends int(time.time()) for this reason. Matching it removes the
only float from anything signed.

The tests now cover the whole listener path over a real announcement, and
record the float hazard so the next person does not rediscover it.

+++ rust/src/announce.rs
@@ -106,7 +106,9 @@ fn run(shared: SharedRef, port: u16, interval: Duration, extra: Vec<String>, key
                 v["mentat_announce"] = secret::SIGNED_VERSION.into();
                 v["boot_id"] = boot_id.clone().into();
                 v["seq"] = seq.fetch_add(1, Ordering::Relaxed).into();
-                v["t"] = secret::now_s().into();
+                // Integer seconds: f64 does not survive a JSON round trip,
+                // so a float here breaks the signature. See secret::canonical.
+                v["t"] = (secret::now_s() as u64).into();
             }
             match &key {
                 Some(k) => secret::sign(&v, k),
